Tashiro manager 0618 2

+++ FarmDefense/Assets/Script/StructCSV.cs
@@ -37,16 +37,16 @@ public class EnemyCSV
 public class WeaponCSV
 {
     public string ID;
-    public float Lv1; 
-    public float Lv2; 
-    public float Lv3; 
-    public float Lv4; 
-    public float Lv5; 
-    public float Lv6;
-    public float Lv7;
-    public float Lv8;
-    public float Lv9;
-    public float Lv10;
+    public int Lv1;
+    public int Lv2;
+    public int Lv3;
+    public int Lv4;
+    public int Lv5;
+    public int Lv6;
+    public int Lv7;
+    public int Lv8;
+    public int Lv9;
+    public int Lv10;
 }
 
 // ƒAƒCƒeƒ€
